Pass the final post IDs as an block attribute
