+++ blocks/query/edit.tsx
@@ -2,7 +2,7 @@
 import { PostPicker, TermSelector, Checkboxes } from '@alleyinteractive/block-editor-tools';
 import classnames from 'classnames';
 import { useDebounce } from '@uidotdev/usehooks';
-import ApiFetch from '@wordpress/api-fetch';
+import apiFetch from '@wordpress/api-fetch';
 import { InnerBlocks, InspectorControls, useBlockProps } from '@wordpress/block-editor';
 import {
   PanelBody,
@@ -18,61 +18,19 @@ import { addQueryArgs } from '@wordpress/url';
 
 import type { WP_REST_API_Post, WP_REST_API_Posts } from 'wp-types';
 import { Template } from '@wordpress/blocks';
+import type {
+  EditProps,
+  Option,
+  Taxonomies,
+  Term,
+  Types,
+} from './types';
 
 import {
   mainDedupe,
 } from '../../services/deduplicate';
 
 import './index.scss';
-
-interface EditProps {
-  attributes: {
-    backfillPosts?: number[];
-    deduplication?: string;
-    maxNumberOfPosts?: number;
-    minNumberOfPosts?: number;
-    numberOfPosts?: number;
-    offset?: number;
-    posts?: any[];
-    query: {
-      [key: string]: string | number | number[] | string[];
-    }
-    postTypes?: string[];
-    searchTerm?: string;
-    terms?: {
-      [key: string]: any[];
-    };
-  };
-  setAttributes: (attributes: any) => void;
-}
-
-interface Taxonomies {
-  [key: string]: {
-    name: string;
-    slug: string;
-    rest_base: string;
-  };
-}
-
-interface Types {
-  [key: string]: {
-    name: string;
-    slug: string;
-    rest_base: string;
-  };
-}
-
-interface Option {
-  label: string;
-  value: string;
-}
-
-interface Term {
-  id: number;
-  title: string;
-  url: string;
-  type: string;
-}
 
 interface Window {
   wpCurateQueryBlock: {
@@ -97,9 +55,9 @@ export default function Edit({
     minNumberOfPosts = 1,
     numberOfPosts = 5,
     offset = 0,
-    posts: manualPosts = [],
+    manualPosts = [],
     postTypes = ['post'],
-    searchTerm,
+    searchTerm = '',
     terms = {},
   },
   setAttributes,
@@ -153,10 +111,7 @@ export default function Edit({
   // Fetch available taxonomies.
   useEffect(() => {
     const fetchTaxonomies = async () => {
-      const path = '/wp/v2/taxonomies';
-      ApiFetch({
-        path,
-      }).then((response) => {
+      apiFetch({ path: '/wp/v2/taxonomies' }).then((response) => {
         setAvailableTaxonomies(response as Taxonomies);
       });
     };
@@ -166,10 +121,7 @@ export default function Edit({
   // Fetch available post types.
   useEffect(() => {
     const fetchTypes = async () => {
-      const path = '/wp/v2/types';
-      ApiFetch({
-        path,
-      }).then((response) => {
+      apiFetch({ path: '/wp/v2/types' }).then((response) => {
         setAvailableTypes(response as Types);
       });
     };
@@ -193,7 +145,7 @@ export default function Edit({
       );
       path += termQueryArgs;
 
-      ApiFetch({
+      apiFetch({
         path,
       }).then((response) => {
         const postIds: number[] = (response as WP_REST_API_Posts).map(
@@ -229,7 +181,7 @@ export default function Edit({
   const setManualPost = (id: number, index: number) => {
     const newManualPosts = [...manualPosts];
     newManualPosts.splice(index, 1, id);
-    setAttributes({ posts: newManualPosts });
+    setAttributes({ manualPosts: newManualPosts });
   };
 
   const setTerms = ((type: string, newTerms: Term[]) => {
@@ -251,7 +203,7 @@ export default function Edit({
   const setNumberOfPosts = (newValue?: number) => {
     setAttributes({
       numberOfPosts: newValue,
-      posts: manualPosts.slice(0, newValue),
+      manualPosts: manualPosts.slice(0, newValue),
     });
   };
 
@@ -260,6 +212,7 @@ export default function Edit({
       manualPosts[i] = null; // eslint-disable-line no-param-reassign
     }
   }
+
   manualPosts = manualPosts.slice(0, numberOfPosts); // eslint-disable-line no-param-reassign
 
   const TEMPLATE: Template[] = [
@@ -294,31 +247,24 @@ export default function Edit({
   return (
     <>
       <div {...useBlockProps()}>
-        <InnerBlocks
-          template={TEMPLATE}
-        />
+        <InnerBlocks template={TEMPLATE} />
       </div>
 
       <InspectorControls>
-        { /* @ts-ignore */ }
         <PanelBody
           title={__('Setup', 'wp-curate')}
           initialOpen
         >
           {minNumberOfPosts !== undefined && minNumberOfPosts !== maxNumberOfPosts ? (
-            <>
-              { /* @ts-ignore */ }
-              <RangeControl
-                label={__('Number of Posts', 'wp-curate')}
-                help={__('The maximum number of posts to show.', 'wp-curate')}
-                value={numberOfPosts}
-                onChange={setNumberOfPosts}
-                min={minNumberOfPosts}
-                max={maxNumberOfPosts}
-              />
-            </>
+            <RangeControl
+              label={__('Number of Posts', 'wp-curate')}
+              help={__('The maximum number of posts to show.', 'wp-curate')}
+              value={numberOfPosts}
+              onChange={setNumberOfPosts}
+              min={minNumberOfPosts}
+              max={maxNumberOfPosts}
+            />
           ) : null}
-          { /* @ts-ignore */ }
           <RangeControl
             label={__('Offset', 'wp-curate')}
             help={__('The number of posts to pass over.', 'wp-curate')}
@@ -334,8 +280,7 @@ export default function Edit({
           initialOpen={false}
           className="manual-posts"
         >
-          {manualPosts.map((post, index) => (
-            /* @ts-ignore */
+          {manualPosts.map((_post, index) => (
             <PanelRow className={classnames(
               'manual-posts__container',
               { 'manual-posts__container--selected': manualPosts[index] },
@@ -346,7 +291,7 @@ export default function Edit({
                 allowedTypes={allowedPostTypes}
                 onReset={() => setManualPost(0, index)}
                 onUpdate={(id: number) => { setManualPost(id, index); }}
-                value={manualPosts[index] ?? 0}
+                value={manualPosts[index] || 0}
                 className="manual-posts__picker"
               />
             </PanelRow>
@@ -368,7 +313,7 @@ export default function Edit({
               <>
                 { /* @ts-ignore */ }
                 <TermSelector
-                  label={availableTaxonomies[taxonomy].name}
+                  label={availableTaxonomies[taxonomy].name || taxonomy}
                   subTypes={[taxonomy]}
                   selected={terms[taxonomy] ?? []}
                   onSelect={(newCategories: Term[]) => setTerms(taxonomy, newCategories)}
@@ -377,11 +322,10 @@ export default function Edit({
               </>
             ))
           ) : null}
-          { /* @ts-ignore */ }
           <TextControl
             label={__('Search Term', 'wp-curate')}
             onChange={(next) => setAttributes({ searchTerm: next })}
-            value={searchTerm as string}
+            value={searchTerm}
           />
         </PanelBody>
       </InspectorControls>
